add WAVECAR.prepareGVEC return type as int
Why:
previously return a ndarray with default dtype

How:
    return np.array(GVEC, dtype=int)

+++ wavefunction.py
@@ -299,7 +299,7 @@ class WAVECAR:
                     if etot < Ec:
                         GVEC.append([igx,igy,igz])
                             
-        return np.array(GVEC)
+        return np.array(GVEC, dtype=int)
 
     def gvectors(self, ikpt=0, gamma=False):
         '''
